requirements.txt
absl-py==0.15.0
anyio==4.1.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==1.0.3
attrs==25.1.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==4.2.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dm-tree==0.1.8
docutils==0.21.2
dominate==2.9.1
etils==1.5.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.0
flatbuffers==1.12
fqdn==1.5.1
fsspec==2025.3.0
gast==0.4.0
gin-config==0.5.0
-e git+https://github.com/KaihangHe/GoChessParse.git@c3342e824767c76d3cae21e306adc6f83dc6939c#egg=GoChessParse
google-api-core==2.10.2
google-api-python-client==2.163.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.34.1
h11==0.14.0
h5py==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==25.6.0
kiwisolver==1.4.7
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.1.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.21
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
referencing==0.36.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.0
rpds-py==0.23.1
rsa==4.9
sacrebleu==2.5.1
scikit-learn==1.6.1
scipy==1.10.1
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
six==1.15.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.16.1
tensorflow-model-optimization==0.7.3
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.18.1
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.5.0
tf-slim==1.1.0
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twine==6.1.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20241206
typing-extensions==3.7.4.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
visitor==0.1.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.0.1
widgetsnbextension==4.0.13
wrapt==1.12.1
WTForms==3.0.0
zipp==3.21.0
